Finish CuratorComponent refactor integration with profile curator Rebase
